pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ht-chpp"
version = "0.1.0"
description = "Modern Python library for Hattrick API data processing"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Tomáš Seidler", email = "tomas.seidler@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.8.0",
    "authlib>=1.2.0",
    "polars>=0.20.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
]

[project.scripts]
ht-chpp = "cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/ht-chpp"
Repository = "https://github.com/yourusername/ht-chpp.git"
Issues = "https://github.com/yourusername/ht-chpp/issues"

[tool.setuptools.packages.find]
include = ["ht_chpp*"]

[tool.setuptools.package-data]
ht_chpp = ["*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100